Refactoring/ adding ideas
Added the enemy database and card database so we can implement new enemies and card in there with ease. This also allows a more widely and easy usage of Enemy class. Battle class is a temporary idea of how to implement turn based with the idea of it being a game state.